♻️ [Refactoring]: #493 lexer.rs を責務ごとに 9 サブモジュールへ分割し facade を 115 行へ縮小
786 行・27 メソッドの lexer.rs を単一責務原則に沿って分散 impl で分割する。
公開 API シグネチャ・可視性・振る舞い・panic 不在契約は完全無変更で、
既存 969 テストは無修正のまま全件パスする。

- cursor.rs / skip.rs / integer.rs / real.rs / name.rs / keyword.rs /
  delimiters.rs / dispatch.rs を新規作成しメソッドを移動
- token_buffer.rs を lookahead.rs へ git mv し、free fn を &mut self メソッド化 +
  peek_token*/take_token* の 5 公開 API を facade から集約
- facade は struct 定義 + new/position/cursor_position/is_eof +
  private hex_value のみを残置

Refs #493
